Moved pipeline related code to its own files
