+++ firemountain/CMakeLists.txt
@@ -31,6 +31,8 @@ target_sources(${PROJECT_NAME}
     ${CMAKE_CURRENT_SOURCE_DIR}/src/vk_mesh.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/vk_init.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/vk_renderer.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/vk_pipeline_builder.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/vk_pipeline.cpp
 )
 
 target_include_directories(${PROJECT_NAME}
