+++ firemountain/include/vk_types.hpp
@@ -3,6 +3,16 @@
 #include <vk_mem_alloc.h>
 #include <glm/glm.hpp>
 
+
+#define VK_CHECK(x)                                                     \
+    do {                                                                \
+		VkResult err = x;                                               \
+		if (err) {                                                      \
+			std::cout <<"Detected Vulkan error: " << err << std::endl;  \
+			abort();                                                    \
+		}                                                               \
+	} while (0)
+
 struct AllocatedBuffer {
     VkBuffer buffer;
     VmaAllocation allocation;
